langgraph/prompt-chaining/pyproject.toml
[project]
name = "prompt-chaining"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Muhammad Fasih", email = "programmer.fasih.18@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.74",
    "litellm>=1.61.9",
    "python-dotenv>=1.0.1",
]

[project.scripts]
prompt-chaining = "prompt_chaining:main"
own = "prompt_chaining.own:main"
llm = "prompt_chaining.llm:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
